refact : 펫플레이스 리팩터링 (#71)

See merge request s07-webmobile1-sub2/S07P12D203!100

+++ Back/aejimeongji/src/main/java/com/ssafy/aejimeongji/api/dto/ScrollResponse.java
@@ -12,5 +12,5 @@ public class ScrollResponse<T> {
     private List<T> data;
     private Boolean hasNext;
     private Long curLastIdx;
-    private Long limit;
+    private int limit;
 }
